test: refactor wallet tests
